pyproject.toml
[project]
name = "JarvisQuestionedBot"
version = "0.1.0"
description = ""
authors = [
    { name = "ebo_shit", email = "isakovn2005@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "aiogram>=3.17.0",
    "jinja2>=3.1.5",
    "pydantic>=2.10.5",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "pytz (>=2025.2,<2026.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "gspread (>=6.2.0,<7.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
]

[tool.poetry.dependencies]


[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
